further pg specific setting up
