+++ spec/db/database.yml
@@ -9,7 +9,7 @@ postgres:
   <<: *default
   adapter: postgresql
   encoding: unicode
-  database: trax_model_test
+  database: trax_model_tests
   pool: 5
   host: localhost
   username: postgres
