+++ spec/support/pg/models.rb
@@ -1,0 +1,5 @@
+module Ecommerce
+  class Product < ::ActiveRecord::Base
+    asdsd
+  end
+end
